requirements.txt
flask
google-cloud-datastore
gunicorn
google-cloud-language
google-cloud-translate
google-cloud-storage
google-cloud-vision
